requirements.txt
darkdetect==0.8.0
ooj==0.1.1
pillow==10.2.0
pystray==0.19.5
pywin32==306